+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+## 1.0.2
+
+- Always match translated ARB files' definition order to the template ARB file.
+- Handle escaped curly brace characters.
+- Handle whitespace that is sometimes introduced by Google Cloud Translate.
+- Provide more information in `InvalidFormatException`
+
 ## 1.0.1
 
 - Fixed bug affecting simple strings that begin & end with a variable.
-- Added InvalidFormatException for mismatched curly braces.
+- Added `InvalidFormatException` for mismatched curly braces.
 
 ## 1.0.0